requirements.txt
aiofiles==23.2.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distlib==0.3.9
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
fb-idb==1.1.7
filelock==3.18.0
grpclib==0.4.7
gunicorn==23.0.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.0.5
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.7
protobuf==5.26.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
six==1.16.0
sqlparse==0.5.1
treelib==1.7.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
virtualenv==20.30.0
